+++ apps/bare-expo/android/app/src/main/java/dev/expo/payments/MainApplication.java
@@ -17,10 +17,8 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import expo.modules.ReactNativeHostWrapper;
 import expo.modules.ApplicationLifecycleDispatcher;
-import expo.modules.devlauncher.DevLauncherController;
 
 public class MainApplication extends Application implements ReactApplication {
-  static final boolean USE_DEV_CLIENT = false;
 
   private final ReactNativeHost mReactNativeHost = new ReactNativeHostWrapper(
     this,
@@ -57,10 +55,6 @@ public class MainApplication extends Application implements ReactApplication {
     super.onCreate();
     SoLoader.init(this, /* native exopackage */ false);
     ReactNativeFlipper.initializeFlipper(this);
-
-    if (USE_DEV_CLIENT) {
-      DevLauncherController.initialize(this, mReactNativeHost);
-    }
     ApplicationLifecycleDispatcher.onApplicationCreate(this);
   }