+++ packages/expo-dev-launcher/android/src/expo-44/java/expo/modules/devlauncher/DevLauncherPackageDelegate.kt
@@ -1,7 +1,11 @@
 package expo.modules.devlauncher
 
+import android.content.Context
 import com.facebook.react.bridge.NativeModule
 import com.facebook.react.bridge.ReactApplicationContext
+import expo.modules.core.interfaces.ApplicationLifecycleListener
+import expo.modules.core.interfaces.ReactActivityHandler
+import expo.modules.core.interfaces.ReactActivityLifecycleListener
 import expo.modules.devlauncher.modules.DevLauncherDevMenuExtensions
 import expo.modules.devlauncher.modules.DevLauncherInternalModule
 import expo.modules.devlauncher.modules.DevLauncherModule
@@ -15,4 +19,8 @@ object DevLauncherPackageDelegate {
       DevLauncherDevMenuExtensions(reactContext),
       DevLauncherAuth(reactContext)
     )
+
+  fun createApplicationLifecycleListeners(context: Context?): List<ApplicationLifecycleListener> = emptyList()
+  fun createReactActivityLifecycleListeners(activityContext: Context?): List<ReactActivityLifecycleListener> = emptyList()
+  fun createReactActivityHandlers(activityContext: Context?): List<ReactActivityHandler> = emptyList()
 }