+++ packages/expo-modules-core/CHANGELOG.md
@@ -9,6 +9,7 @@
 ### 🎉 New features
 
 - Add `getDevSupportManagerFactory` support to `ReactNativeHostHandler`. ([#16434](https://github.com/expo/expo/pull/16434) by [@lukmccall](https://github.com/lukmccall))
+- Add support for automatic setup of `expo-dev-client` on Android. ([#16441](https://github.com/expo/expo/pull/16441) by [@esamelson](https://github.com/esamelson))
 
 ### 🐛 Bug fixes